move patterns to separate file and replace rng with a pre-sorted random number array
